requirements.txt
absl-py==0.11.0
astor==0.8.1
cached-property==1.5.2
click==7.1.2
Flask==1.1.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.33.2
gunicorn==20.0.4
h5py==3.0.0
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MarkupSafe==1.1.1
numpy==1.19.4
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
six==1.15.0
tensorboard==1.15.0
tensorflow==1.15.3
tensorflow-estimator==1.15.1
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
